修正 FID/KID 输入类型为 uint8
_to_255 现将图像转换为 torch.uint8，符合 torch_fidelity 的 FID/KID 输入要求

避免 float 输入导致的 ValueError: Expecting image as torch.uint8

+++ utils/metrics.py
@@ -69,7 +69,7 @@ def _as_rgb(tensor: torch.Tensor) -> torch.Tensor:
 
 
 def _to_255(tensor: torch.Tensor) -> torch.Tensor:
-    return (tensor * 255.0).clamp(0.0, 255.0)
+    return (tensor * 255.0).clamp(0.0, 255.0).to(torch.uint8)
 
 
 def _to_lpips_range(tensor: torch.Tensor) -> torch.Tensor:
